# `<regex>`: Avoid allocating during matching when regex defines no captures